+++ changelogutils/changelogdocutils/merged_release.go
@@ -4,11 +4,7 @@ import (
 	"context"
 	"fmt"
 	"github.com/google/go-github/v32/github"
-	"github.com/rotisserie/eris"
 	. "github.com/solo-io/go-utils/versionutils"
-	"io/ioutil"
-	"net/http"
-	"regexp"
 )
 
 type DependencyFn func(*Version) (*Version, error)
@@ -20,38 +16,6 @@ type MergedReleaseGenerator struct {
 	openSourceRepo       string
 	releaseDepMap        map[Version]*Version
 	FindDependentVersion DependencyFn
-}
-
-// Default FindDependentVersionFn (used for Gloo Edge)
-func FindDependentVersionFn(enterpriseVersion *Version) (*Version, error) {
-	versionTag := enterpriseVersion.String()
-	dependencyUrl := fmt.Sprintf("https://storage.googleapis.com/gloo-ee-dependencies/%s/dependencies", versionTag[1:])
-	request, err := http.NewRequest("GET", dependencyUrl, nil)
-	if err != nil {
-		return nil, err
-	}
-	response, err := http.DefaultClient.Do(request)
-	if err != nil {
-		return nil, err
-	}
-	body, err := ioutil.ReadAll(response.Body)
-	if err != nil {
-		return nil, err
-	}
-	re, err := regexp.Compile(`.*gloo.*(v.*)`)
-	if err != nil {
-		return nil, err
-	}
-	matches := re.FindStringSubmatch(string(body))
-	if len(matches) != 2 {
-		return nil, eris.Errorf("unable to get gloo dependency for gloo enterprise version %s\n response from google storage API: %s", versionTag, string(body))
-	}
-	glooVersionTag := matches[1]
-	version, err := ParseVersion(glooVersionTag)
-	if err != nil {
-		return nil, err
-	}
-	return version, nil
 }
 
 func NewMergedReleaseGenerator(client *github.Client, repoOwner, enterpriseRepo, ossRepo string, dependencyFunc DependencyFn) *MergedReleaseGenerator {
